Default the search path to empty for config file parsing
Summary: I want to add support for src layouts (see next diff), for which I want to add src/ rather than ./ as the guessed search path. This is easier if the search path starts empty and we don't add a guess until we're at a point where we can figure out the project layout.

Reviewed By: connernilsen

Differential Revision: D74163799

fbshipit-source-id: 8f7092d0931e3269b373a0bac46660acff6d1b9c