Support `pyre-ignore-all-errors`
Summary:
I'm not sure we want to support this long-term, but it seems like for
pre-release testing and also as the rollout happens it might be convenient
to understand existing pyre-style ignore-all directives.

It's easy to complain about these in a linter and remove support once
pyre directives have been cleaned up.

Reviewed By: grievejia

Differential Revision: D74181043

fbshipit-source-id: 1b246955e07b48def03e5522056abb81ba3cecf7