Add support for top-level `# type: ignore`
Summary:
Our implementation follows the spec as in
https://typing.python.org/en/latest/spec/directives.html#type-ignore-comments
but with the additional restriction that the line immediately following a
`# type: ignore` must be blank or a comment.

The restriction allows us to continue supporting the current behavior (adapted
from `# pyre: ignore`) where a line-level suppression can be provided on the
preceding line; this technically violates the spec (which says the suppression
can only be applied on the line with the comment) and may be worth revisiting
now that we have a `# pyrefly: ignore`, but I wanted to get basic support
for `# type: ignore` working without having to simultaneously discuss changing
`# type: ignore` behavior.

(+1 conformance from this, since top-level `# type: ignore` is in the spec, and the conformance
test meets our heuristics even though our implementation isn't quite for-the-letter lined up
with the spec)

Reviewed By: grievejia

Differential Revision: D74181047

fbshipit-source-id: e36320ed4eadeabf95c79fca936cbe75cafcb4d1