Add a test of an attr set in an unrecognized method
Summary:
This diff begins a stack where I will infer attribute types for attributes
defined anywhere in a class; we will error at the definition if it is defined
in an unrecognized method ("recognized" means a constructor or a recognized
test setup method in this context; we could potentially extend the notion
of recognized methods later), but attribute accesses in downstream code ought
to be analyzed correctly.

We might already have a test of this, but it's a little hard to search tests
systematically and I need a fixture to develop against.

Reviewed By: grievejia

Differential Revision: D74272473

fbshipit-source-id: c44c626f0a41325b18a9e3ca4ca980e503e0708a