Use an API for adding method-defined attrs
Summary:
I'm starting work with this diff to support inferring attribute types
for attributes not defined in constructors and other "supported" methods
like test setup methods.

Our intention is to throw a type error (which might wind up being default suppressed,
but we at least want it for our most strictness-loving users) on the *definition* site,
but we think the current behavior of pretending the attr doesn't exist is not compatible
with our goal of being flexible and IDE-first.

I looked at the implementation and realized that unless we want to use an
inefficient two-pass loop iterating over the methods, we really need to
determine at attribute identification time (at the end of a method body in
bindings) whether the method is a "supported" constructor / test setup method
or not; otherwise it will be tricky to give proper precedence.

This diff begins moving in that direction by abstracting the code for recording attributes
so that we call a helper method instead of directly mutating raw data.

Reviewed By: grievejia

Differential Revision: D74272469

fbshipit-source-id: 91209c4ef54854e36d05256b38d593dd9b9c6e53