Support attributes defined outside constructors
Summary:
In this diff we finally use the flag we wired up to create a ClassField for
an attribute defined outside of a recognized method, but also produce a type
error.

For default Pyright-like behavior, users could now ignore that error in their
config to treat all implicitly defined attributes as valid.

There's a regression in the conformance result; I think it is arguably a bug in
the test for which we should consider putting a PR up - we're supposed to complain
when a protocol sets a `self` attribute in a method which makes sense, but the
test is also currently verifying that we don't (in the face of that error) consider
the protocol to have the attribute thus defined which seems overly proscriptive to me;
the regression happens because as of this diff we do in fact detect the attribute
(`self.temp` defined implicitly in `Template.method`.

Reviewed By: grievejia

Differential Revision: D74272470

fbshipit-source-id: 94bf2678041d361c20ef8bad41ae48fe5a8d9a11