Push more logic into the ClassBody datatype
Summary:
Given the current logic, there's no reason for us to be saving attributes
defined outside of recognized methods at all (that will change soon, but
in meantime it's clearer to clean up the existing logic).

We can do the filtering there, and also dump out a flat iterator of the
method_defined_attributes instead of exposing the raw data of a two-level
map.

These two helpers will give us a clear place to track when an attribute is
defined by a method not recognized as a valid attribute-defining method.

Reviewed By: grievejia

Differential Revision: D74272472

fbshipit-source-id: 20bf2e83f93ac0673f45c8b207e074eac7ead8f8