Allow passing both a config file and files to check on the same command line
Summary:
We discussed this in an eng sync a couple weeks ago, and the conclusion at the time was to wait a bit because Neil was refactoring config finding in a way that may make this easier in the future.

The future is now, and implementation-wise, this has indeed become pretty simple.

I think we should support this because it's the obvious way to type-check a set of files with a known configuration. Neither of the alternatives is great:

* Omitting -c to let the config be auto-discovered works most of the time but is less explicit. It's also sort of odd that there's something pyrefly allows you to do but doesn't allow you to explicitly request.
* Editing the project_includes in the config file always works but is tedious.

As I understand it, the main argument against this is that it allows you to type-check files with configurations whose project_includes don't cover them. This is already implicitly allowed today (if we find a config through upward search, we use it regardless of what its project_includes is). This diff would make it possible to type-check a file with a config on a totally unrelated path, which is a bit questionable, but IMO it's okay to allow this because it's very hard to do *accidentally*. You have to explicitly request both the config and the file, and at that point, you probably know what you're doing.

Reviewed By: grievejia

Differential Revision: D74207311

fbshipit-source-id: 25e7edae2e91661909560f9d35c6355cab330b94