Handle output paths better
Summary:
First, enforce that supplied output_paths must end in pyrefly.toml or pyproject.toml file. This is to avoid the difficulty of determining if a random path is a directory or file and if it exists. (e.g. tmpdir() gives us paths like `/tmp/.tmpABC`, which looks like a file to Rust.) It is not onerous to force the user to disambiguate.

Second, create the parent dirs of the output path if they don't exist. This is a "do the right thing if it's obvious" thing.

Reviewed By: stroxler

Differential Revision: D73860743

fbshipit-source-id: b46ccd2973f8a1188c66c192cf04636e8072f7e1