Print errors to stdout instead of stderr
Summary:
The previous diff on the stack makes it such that we print to `stdout` when `--output-format` is set to `json`, but prints to `stderr` when `--output-format` is set to `text`. This diff align these setups and have text errors prints to stdout as well.

Fully fixes https://github.com/facebook/pyrefly/issues/157

Reviewed By: connernilsen

Differential Revision: D74293032

fbshipit-source-id: a46f6c501a5b6c2d7bf6473fda5c1d01a392e17f