Move environment into new environment module
Summary:
I'm going to be making some new functionality (before and after PyCon) to automatically figure out the PythonEnvironment we should use when not explicitly provided. This diff creates the new `config::environment` module, which is where we'll place all of this new functionality.

Upcoming in the stack, we'll add functionality for:
- finding an interpreter in a venv
- using the venv finder when searching for a python environment before falling back to the first interpreter available on the path

Reviewed By: rchen152

Differential Revision: D74363908

fbshipit-source-id: 1403d2cdb46cf111e81a3826e7e62b08f12bb243