tuple length narrowing
Summary:
We can narrow tuple types based on length. This is an optional feature per the conformance tests.

len(x) == y will eliminate concrete/named tuples that definitely do not match the length, and turn unbounded tuples into concrete tuples with the matching length. for unpacked tuples (concrete prefix + unbounded middle + concrete suffix), if the prefix + suffix are too long then we eliminate it as well, otherwise we make a concrete tuple with as many middle elements as necessary.

len(x) != y will eliminate concrete/named tuples that match the length

This does not work in match statements yet, the narrowing there is buggy and needs to be refactored before it can be added.

Reviewed By: stroxler

Differential Revision: D74329909

fbshipit-source-id: 7b00d85ae406e5888328ce0fc38840eecd6fc2cc