Switch to a mostly indexing based global find references
Summary:
It turns out that using a recheck based find references can be quite expensive. Although rechecks are generally fast, we have to recheck under `Require::Everything` to obtain various traces, which is way more expensive than a usual recheck (up to 2x), and used about 9x more memory.

This diff switches find-refs to use an indexing based one. You might think that it might consume a lot more memory, because there are too many references in the code. However, note that we don't actually need to store every single reference. We only need to restore references that will jump to an **external** definition.

Why we don't need to store internal references? Consider the case where we want to find references of an exported symbol in `foo.py`. After finding all the local references within `foo.py`, we need to scan the index to find all the external references. None of the local-references in other files will contribute to the solutions, so there is no point storing them. As a result, the new index only introduces minimal memory overhead.

The indexing step should be very fast as well, since we only need to scan through the bindings once to find all the external variable reference, and find the definition location for all attribute during checking (which could be made even faster by making the type checking API also return def-loc, but not done in this diff).

We will continue to use `Require::Everything` to find definitions within the definition file, but we only need to upgrade the checking for one file, so it should be pretty fast.

Reviewed By: ndmitchell

Differential Revision: D74155343

fbshipit-source-id: 1a87a5a9b6f8e393a31e46d3f8ae5eb4ada5ea7e