Extract function header traversal
Summary:
There's a lot that goes on in function header traversal, but there are actually
only two values we extract in the end: the map of legacy type parameters and the
return annotation key + text range.

Pulling this out shrinks `function_def` quite a bit, and gets rid of a lot of
local variables like `tparams` that looked important (and were in scope for
over 100 lines) but were actually local to this little bit of logic. And it makes
it much easier to see the actual results of the traversal depended upon elsewhere.

Reviewed By: kinto0

Differential Revision: D74372516

fbshipit-source-id: 9ee495619d2518a2a76543c966dec17edd38a75d