Reorganize some code
Summary:
I'm still trying to group `function_def` into logical units I can wrap my brain
around, and one challenge remains that the blocks of code are not grouped in
a clear way. Here, I am collecting a bunch of metadata definitions that are global
to the function in question and don't depend on anything about the actual code so that
they all live above the code analysis, which makes the dependency clearer.

Reviewed By: kinto0

Differential Revision: D74372506

fbshipit-source-id: e59e9b71fb87c5ddaedec7a84fa7f6080e282412