Extract `return_annotation_with_range` helper
Summary:
I am finding that the `function_def` code is too large for me to make refactors
with the ease I would like; some of the problem is that it's just too hard to
scroll, and some of it is that there are so many variables in scope it is hard to
track them.

So I'm going to start pulling out bits where I encounter confusion during a refactor;
in this case I stumbled upon confustion between `return_annotation` which is an expression
and `return_ann` which is an optional `Idx<KeyAnnotation>`; using a helper means we don't
have to have both these names in the same scope.

Reviewed By: kinto0

Differential Revision: D74372512

fbshipit-source-id: 7976d8cbab603a74e22d75ab97704a91cad4195e