Rename `return_annot_range`
Summary:
The name "return_annot_range" means, to me "the range fo the return annotation".

That's not what this variable is, it's an option of the annotation *with* the
range.

I got confused about this and wasted some time in a refactor, figured I should
just change the name so no one else does.

Reviewed By: kinto0

Differential Revision: D74372508

fbshipit-source-id: 89d52a10531ab35f6e266ed898e26a418ff88708