More regrouping and simplification
Summary:
The biggest single problem with `function_def` from the standpoint of
refactors remains the sheer number of local vars in scope, which is both
tied to the total number of vars and the fact that some of them live for
quite a while before use.

Here I regroup code so that `implicit_return` is defined in a single block
of code that does not require any variables escaping into the top-level
function namespace.

This also makes it clearer that `implicit_return` does *not* depend on
the `self.stmts` call that traverses the body; in the previous state,
this was already the case but was hard to see, since `self.stmts`
does a lot of mutation and there's really no way to be sure from just
a skim without already knowing all the details.

Reviewed By: kinto0

Differential Revision: D74372513

fbshipit-source-id: 37d861111175e0cac463e0c07efb4d56a7b273e7