Reorganize some function header logic to be more tightly grouped
Summary:
The ordering of things in `function_def` is legal because lots of unit tests
stress it, but in some cases it doesn't actually make sense - code that belongs
together has had unrelated logic interspersed in the middle of it.

This is just a minor reorganization so that function header code is more clearly
grouped.

Reviewed By: kinto0

Differential Revision: D74372507

fbshipit-source-id: 0f48f6394aafc393e6d842ffc9271b1041a64950