Rename yield and return accumulators
Summary:
Throughout the bindings code, the yield-and-return-accumulation variables and
fields seem to have less clear names than I would prefer - we call the struct
itself `FuncInfo`, the stack of yield-and-return structs `self.functions`, and
the individual value popped off for a given function `accumulator`.

I think emphasizing in all of the names that these are yield-and-return
collectors makes it a lot easier to see what's happening, especially if you
stumble across this code in a context where you're not already thinking about
yield and return collection.

Reviewed By: kinto0

Differential Revision: D74372517

fbshipit-source-id: 89b7717d795f0f8862f714dbf6161dbeb75249ba