Tweak the handling of self attr assignments for methods
Summary:
Instead of passing a raw `Scope` out from `function_body`, which could
potentially be used for all kinds of things, let's pass out an optional
`MethodInner`.

This makes it a lot more obvious at a glance that this data is only needed.
for methods, and by naming the variables consistently we also make it
pretty easy for a code search to show where and how it gets used.

Reviewed By: kinto0

Differential Revision: D74372511

fbshipit-source-id: 71cfb77e3763a7605c056806cf2d004247277370