Make some logic clearer
Summary:
From the point of view of trying to group function_def code into logical blocks,
it would be nice if we could handle popping the annotation scope below all of the
return type inference logic.

We probably can, but trying to do this naively leads to problems because if we move
the pop below the block that tracks `self.<attr>` assignments, then the `if let` will
always silently fail and lots of things break.

This isn't so easy to see the way the code was laid out, with no comments about *what*
is getting popped and the pop quietly just stuck right after `func_scope = self.scopes.pop()`;
I spent a while trying to figure out why a seemingly trivial refactor broke things.

This diff just adds a comment explaining (a) what is being popped and (b) why it needs
to be popped at exactly this point in control flow, while also putting the pop next
to the use of `self.scopes.current_mut()` that depends on it instead of next to the
:x

Reviewed By: kinto0

Differential Revision: D74372515

fbshipit-source-id: be2ff7519f142968a4b3a6361088bfe1d7e1b383