Rearrange some code in `function_def`
Summary:
I want to extract the function body traversal into a helper, which will be a
big help in getting to the point where I can implement a mode where we skip
return type inference and type checking function bodies.

This diffs prepares for that by rearranging the innermost function-body-traversal
logic to be in a contiguous block.

Reviewed By: kinto0

Differential Revision: D74372505

fbshipit-source-id: 4106208737510bfc74e55bf010070e929c38d574