Rename `function_source` / `source` / `FunctionSource`
Summary:
As I was refactoring, I kept getting confused about what this was, it sounds
like it would be the source code of a function, where as it's really just a
two-variant enum: is this a stubbed function, or one with an implementation?

I think using a more descriptive type and consistent variable / field names
helps, then it's obvious.

Reviewed By: kinto0

Differential Revision: D74372510

fbshipit-source-id: 3177278dd6081ec8a7b0104d7eadfe0dde44c7d2