Extract function body analysis into a helper
Summary:
The full body of logic for function body analysis includes not only the function
scope traversal that we already extracted, but also return type inference (including
the handling of implicit returns, which involves a separate pass over the body to
find the last expression(s) in control flow).

This is the unit of work that has to change for us to support a PEP 484 compatible
mode where functions that lack annotations are not analyzed and the return type is
treated as `Any`.

This is also the first stage at which I can actually read the `function_def`
function and understand what is going on without lots of scrolling... it's now
small enough to, for example, easily see the corrrespondence between pushing
annotation scope vs popping it to get back to the class body scope when this is
a method.

Reviewed By: kinto0

Differential Revision: D74372518

fbshipit-source-id: 91d10e88b3f8c31beda2aaa6259553e74f485ce2