Having a specialized event loop to process different tiers of ServerEvents
Summary:
We do have code to skip processing canceled requests, but right now it's completely useless. Right now we are processing all the requests serially, and the cancellation notification arrives after the request, so we will never see the canceled request id being added to the `canceled_requests` set.

To resolve this issue, and to unlock future optimization we can do wrt to the scheduling of request handling, I will setup a specialized event loop. For now, we will have two separate channels:

1. Priority events
2. Normal events that can be further queued

We will use the selector from `crossbeam_channel` crate to favor processing priority events first. For now, we keep most of the other events in the queued events channel, and only have the `RecheckFinished` event that's previously "immediately handled" and the new `CancelRequest` event (so that the server has a chance to know that certain requests can be canceled).

Reviewed By: kinto0

Differential Revision: D74014749

fbshipit-source-id: db702f9aebdaf71237f14b3b3f2200c3d136e0a7