Change ConfigFile::from_file to return (ConfigFile, Vec<Error>)
Summary:
Currently, errors in ConfigFile::from_file are sometimes fatal and sometimes not, depending on whether the config file is passed in on the command line or automatically discovered. IMO it's better for the errors to be non-fatal:
* It lets us keep going and show as much useful info to the user as we can.
* It will allow us to simplify our error reporting scheme, which currently has to jump through some hoops (warnings, boolean flags) to avoid exiting early on low-severity/recoverable errors.

This diff takes a first step toward making errors non-fatal by changing from_file to return a ConfigFile and a list of errors. For now, callers just emulate the previous behavior by returning the first error when one exists.

Reviewed By: ndmitchell

Differential Revision: D74375303

fbshipit-source-id: 9b65d327f21e0b8a28eb872def348f7d17b328a5