Separate assertions from calculations in `Definitions` tests
Summary:
codecleanupraccoon

In tests of `definitions.rs`, there's this big `check()` function that both calculates the definitions of a module, and assert certain stuffs on the calculation results. Then, the function returns the result definitions, only for the caller side to perform additional asserts on that result...

I think it would be cleaner if we just have a function that calculates the results, and separately perform all the assertions on the caller side. This way, we make it possible to write tests (in stacked diffs) that have more freedom to choose what it wants to assert, as opposed to having to perform the same asserts that were hard-baked into `check()`.

Reviewed By: stroxler

Differential Revision: D74370554

fbshipit-source-id: 9191f2f2871bdbb9cd1cb75126098db1c5824a6e