Always error on extra fields in configs
Summary: Changes ConfigFile::from_file to always error on extras, rather than having a parameter to control the behavior. The parameter made sense when errors in from_file would cause pyrefly to fall all the way back to a default config, since that would be undesirable when the parsed config is otherwise usable. But now we use the parsed config regardless, so we might as well report these errors always.

Reviewed By: connernilsen

Differential Revision: D74380983

fbshipit-source-id: 9357eb0161b3d3d4f5cf7858df9b4d9ecb67ccdc