Always return a ConfigFile from "load" step of config finding
Summary:
This means that we always use the config that ConfigFile::from_file returns, even if we also got some errors,, and that we report all the errors rather than just the first one.

This also fixes an issue we had with the INFO message we print for full-project checks - if the config file didn't parse, we'd say we were using a default config, which is confusing. Now we correctly report that we're analyzing the project configured at <config path>, even if the config contains errors.

Reviewed By: connernilsen

Differential Revision: D74378043

fbshipit-source-id: 57c33c842be7e890c2740e758da227a99ccbb453