Fix bug in async function return type inference
Summary:
We were double-counting the `is_async` flag - we were wrapping the type
inside of `Binding::ReturnType` (but inconsistently - in the case where
we had an annotation, we were producing the annotation type without the
awaitable wrapper) and then wrapping it *again* inside of
`Binding::FunctionDef`.

This diff makes us only wrap in `Binding::FunctionDef`, so that the semantics
are that `Binding::ReturnType` always produces a type as it would be written
in a return annotation - without the awaitable wrapper - unless the function is
an `AsyncGenerator` (in which case the return annotation would have to be
explicitly async, so the semantics are still consistent although Python's behavior
here is a bit confusing in my opinion).

Reviewed By: rchen152

Differential Revision: D74420059

fbshipit-source-id: 677a640c8d1f083d16a7d73f73a5fac4d226a724