Move return type inference into a block
Summary:
This is a trivial code rearrangement, but really helps me visualize how
the code needs to work once we have modes that can (a) either check a function
body without doing return type inference or (b) skip the check entirely.

The return-related data is only needed when we are inferring return types, so
this will work with a branch where the nested block is now, where in all branches
we need to produce `self_assignments` information.

Reviewed By: yangdanny97

Differential Revision: D74441071

fbshipit-source-id: 8afeb721b90c055c06651df36a4fb0c9e27fcf18