Make a custom datatype for `SelfAssignments`
Summary:
After a quick conversation with Niel, we think that it's going to be important
to infer `self.<attr>` existence (with the `Any` type) for attributes defined in
untyped methods.

This means we need to decouple the logic used to save this from the `MethodInner`
datatype, which is tightly tied to full analysis of a scope that won't happen (instead,
we'll use a simple visitor in the untyped case).

This diff just wires up a new datatype we can use for this.

Reviewed By: yangdanny97

Differential Revision: D74441075

fbshipit-source-id: 81c0314555267cd19a415625ff020d1d8eae073d