Push `self_assignments` logic down to `function_body_scope`
Summary:
Now that other things have been cleaned up more, it was jumping out to me
that passing a raw `Scope` up from `function_body_scope` to `function_body`
is mixing abstraction levels (and doing so in a way where the types don't
give us perfect information, since we have to match on the scope kind).

It's significantly cleaner to handle the scope matching and create the
`self_assignments` data inside the lower-level helper; this logic won't
run if we skip checking the function body (in which case we'll need an alternate
way to produce `self_assignments` with gradual types).

Reviewed By: yangdanny97

Differential Revision: D74441073

fbshipit-source-id: 1d5984e5936b4a4b5be9b5cbc11164ff6a2af00d