Add suppression tests (3/n)
Summary:
Our current upgrade tooling avoids adding error suppressions to unparse-able files. We might want to change this in the future, but for now I think it's a decent practice.

Diff also adds a helper function for cases where we want to confirm an error was thrown, etc. during suppression

Reviewed By: yangdanny97

Differential Revision: D74438177

fbshipit-source-id: 066aba6d7bb3fad0e31c1452cb771f2f67dbd1f1