walk to find interpreters
Summary:
This diff adds a `walk_interpreter()` function to search for interpreters within a given directory up to a max depth. It will be used in an upcoming diff by a `Venv` `Finder` to search for possible venv interpreters.

I'm leaving it in the `finder` module for now, since I think it may end up being useful for other finders in the future.

Upcoming in the stack, we'll add functionality for:
- finding an interpreter in a venv
- using the venv finder when searching for a python environment before falling back to the first interpreter available on the path

Reviewed By: stroxler

Differential Revision: D74367600

fbshipit-source-id: 34230dc92644d097dca0a9fbe9f0fcbbeedde5d3