Create venv implementation of Finder
Summary:
This diff creates a `Venv` implementation of `Finder`, which passes in the project directory and searches for a possible interpreter.

The strategy we use for this is similar to [vscode-python](https://github.com/microsoft/vscode-python/blob/b44b4d442fce5c0ab1547c777b846c1ade889832/src/client/pythonEnvironments/common/environmentManagers/simplevirtualenvs.ts#L14), which given an interpreter path, looks for a `pyenv.cfg` file in the same or parent directory of the interpreter. It is preferred to take the more deeply nested interpreter according to vscode-python.

Example: if we have `.venv/pyenv.cfg`, valid an interpreter is expected to be found at either `.venv/bin/python(\d(\.\d)?)?` or `.venv/python(\d(\.\d)?)?`.

Upcoming in the stack, we'll add functionality for:
- using the venv finder when searching for a python environment before falling back to the first interpreter available on the path

Reviewed By: stroxler

Differential Revision: D74363893

fbshipit-source-id: 4a2c9959b754ce51f601ca86eccc5eeaf4150534