Use new PythonEnvironment::find_interpreter() function
Summary:
This diff finishes the functionality required to get venv working without having an interpreter sourced.

In this diff, we switch to the new `PythonEnvironment::find_interpreter()` function when an interpreter is not provided to search for the interpreter to use. This replaces the call to get the default interpreter, which is used as a final fallback in the `find_interpreter()` function instead.

Reviewed By: stroxler

Differential Revision: D74366025

fbshipit-source-id: 084e06f49458d25178d382f0adf1d3228fbd361d