Add find_interpreter function
Summary:
This diff creates a new `PythonEnvironment::find_interpreter()` function, which we'll use in upcoming diffs to find an interpreter for the given project path. This function's current strategy is to try finding a venv, if we're given a project path, and otherwise falls back to the default of `PythonInterpreter::get_default_interpreter()` if one cannot be found.

In the future, we can expand on this strategy to add more logic/finders for different environments that we'll support.

Upcoming in the stack, we'll add functionality for:
- using the venv finder when searching for a python environment before falling back to the first interpreter available on the path

Reviewed By: stroxler

Differential Revision: D74363907

fbshipit-source-id: ece3b401548dffa2e8c8a96dfcfb6b02495e1535