Create Finder trait for searching for Python interpreters
Summary:
This diff creates the `Finder` trait. I'm not sure if a trait is the right way to go yet (over an `impl Fn(&Path) -> Option<PathBuf>`), but I'm going with this approach for now because it seems like it could be useful. If, in the future, we decide a trait isn't worth doing, then we can always go back to the other approach.

The finder trait represents an interface for finding different kinds of Python interpreters and environments based off the path of the project we're configuring. Currently, we're just looking at `venv`, but in the future, I'd like to expand this functionality to deal with conda, pyenv, etc., and more closely model the approach that [vscode-python](https://github.com/microsoft/vscode-python/blob/main/src/client/pythonEnvironments/index.ts#L174) takes.

Upcoming in the stack, we'll add functionality for:
- finding an interpreter in a venv
- using the venv finder when searching for a python environment before falling back to the first interpreter available on the path

Reviewed By: stroxler

Differential Revision: D74366026

fbshipit-source-id: 6a2a01e328066672808b8a7cbb2522859e67913b