Run build_binaries workflow on every push
Summary:
I had previously specified that the workflow would run only on changes to key files, but this had the unfortunate side effect that it's hard to tell when the workflow failed (because the build would go red, then green again).

Just run on every change, so the build stays red if build_binaries is failing.

Reviewed By: yangdanny97

Differential Revision: D74492390

fbshipit-source-id: 4ac59d9ec024d3c3aeeb575d3b497279515b2d81