Support goto definition of attribute of imported module
Summary:
In this diff, for attributes of imported module, we will now report its definition as `AttrDefinition::PartiallyResolvedImportedModuleAttribute(imported_module_name)` instead of `None`. Later in `lsp.rs` where we have access of transaction, we then resolve it to `TextRangeWithModuleInfo`. In this way, go-to-definition can work on this kind of attributes.

For reference indexing, we can just store it into `externally_defined_variable_references`, since the following two kinds of references are equivalent:

```
import foo
foo.bar

# equivalent to
from foo import bar
```

Reviewed By: kinto0

Differential Revision: D74487779

fbshipit-source-id: f4efa06b76e68627dbdaad056608233ad353eb55