Fix colors for landing page
Summary:
This diff updates the colors for the landing page + navbar + docs. I've updated the light theme to use a light grey and the dark theme to use a very dark gray.
The navbar and footer were also modified to have the same color as the background, which I believe looks much nicer.

There are still a couple things that needs to be fixed
- e.g. border around code blocks
- modify icons to be a more visible color for dark theme

Reviewed By: yangdanny97

Differential Revision: D74355413

fbshipit-source-id: 965ff773899558cc5495c369342140940fadbcc5