move code snippet styles to bottom
Summary:
I'm going to make changes to the styling here in D74355413, but I noticed that the stylesheet was at the top of this file instead of the bottom. The current code styling should always include the stylesheet at the bottom so that it's easier to see the file content at the top.

This diff simply moves the stylesheet down to the bottom of the file

Reviewed By: stroxler

Differential Revision: D74418149

fbshipit-source-id: 22a834594146e39bc5e964a14a16ab471e2f41df