move discord__link into custom.css
Summary:
I was looking at `custom.css` and noticed that we included the details for `discord__link` in `code-block-buttons.css` instead. There's a comment saying `/* Custom CSS for the discord logo - for some reason does not work from custom.css */`, but I tested this again and it seem to work just fine when moving it to custom.css.

Moving this so that all icon css is in one place.

Reviewed By: grievejia

Differential Revision: D74343838

fbshipit-source-id: 1781e0e8ffcc80a40a502c29330d123bcd797814