build extensions on ubuntu 20.04 to support older glibc
Summary:
Older operating systems have an older glibc version that won't work with newer versions. But newer versions always work with older version. When a version is too old, our binary fails to start with a message like `version `GLIBC_2.39' not found` (User report [here](https://github.com/facebook/pyrefly/issues/163)).

Ubuntu 22.04 is the oldest version [github will support by default](https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories). By building on ubuntu-latest, [we build on ubuntu 24.04](https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories). But this does not work with older operating systems like ubuntu 22.04 or CentOS 9. Ubuntu 20.04 is EOL end of may 2025 and no longer receives security updates. But CentOS 9 gets updates until 2027. If we build on ubuntu 20.04, we can target CentOS 9 and other older systems.

We have a few options in resolving this:
- build with musl for a statically-linked self-contained binary. Downsides: [known perf regressions](https://github.com/rust-lang/rust-analyzer/issues/16793#issuecomment-2016396816) (I have not benchmarked this). All OS versions (even new ones) will get the `musl` build because there is no way to target older OSes in VSCode's release
- build with an older OS github supports natively like ubuntu 22.04
- build with an even older OS with a custom container

This diff implements option 3 for the extension build. If we like it, we can also use it on the pypi build.

Is it worth it? Should we just target 22.04+? Or should we build with musl?

Reviewed By: SamChou19815

Differential Revision: D74490281

fbshipit-source-id: c904d2bff7e64b2527ef1839cff8a2cd77dc4724