e2e test on all extension builds
Summary:
we build the extension on every push. we should also e2e test it.

this e2e test reuses the build, grabs the binary out of it, and runs an e2e test using that binary.

it's a little complicated. one way to make it simpler is to separate build_extension from build_extension_binary (then we wouldn't have to extract it to get the binary). but this has other drawbacks:
- D74490281 complicates the build and requires a custom OS for linux. i'd rather keep that hack confined to one file

I'm open to future improvement.

Reviewed By: SamChou19815

Differential Revision: D74504276

fbshipit-source-id: 3efb4f155646c40d5a3293a08005fa6ceed7ef6f