Update favicon if in dark mode
Summary:
Similar to D74427257, the favicon doesn't render well when the browser is in dark mode. There doesn't seem to be a setting in docusaurus config for dark mode favicon so I need to modify the svg directly. The browser seem to cache the favicon so it's hard to test this, putting this change up and we can test after land

 {F1977844115}

Reviewed By: stroxler

Differential Revision: D74508340

fbshipit-source-id: 0e54403f140b8b0728fe772d6b6fbca615f48946