always default to dark mode irrespective of preferred color scheme
Summary:
The website looks much better in dark mode than light mode. Setting this as the default. The user still has the option to swap to light mode if they wish with the dark mode toggle.
The docs indicate that `respectPrefersColorScheme` is the right flag to change: https://docusaurus.io/docs/api/themes/configuration

Reviewed By: stroxler

Differential Revision: D74505713

fbshipit-source-id: 1a022d73717935d1049e8903573c6109643705d2