Implement infer-returns-any logic
Summary:
It turns out that it's relatively easy at this stage to implement a mode
approximating mypy with the check_untyped_defs, where we always check a
function body but infer `Any` as the return type when there is no
return annotation.

For the moment I am just using the same logic when we are supposed to skip
untyped functions, because skipping is slightly harder: we need to implement
a visitor to collect `self.<attr>` assignments or it might cause more problems
than it solves.

Note that at the moment we unconditionally produce `Any` as the type; after
talking it over with Neil I decided this is the best thing to do at first because
it is simple, but we may want to consider a simple analysis of `async` and `yield`
keywords to guess slightly more informative types. I left a TODO as a reminder
to consider this improvement.

I'll work on proper skipping next, but in the meantime we can at least mimic
the behavior for downstream code by inferring `Any` as the return type.

Reviewed By: yangdanny97

Differential Revision: D74441074

fbshipit-source-id: 129fccd8b4c3f994092410638f828c5ead70ce5d