Thread `skip_untyped_functions` into `Context`
Summary:
This diff begins a stack implementing a PEP-484-compatible (and mypy-defaults compatible)
mode where we skip untyped definitions and infer a return type of `Any`

The first step is to start propagating the config flag we already support, which eventually
needs to make it into the bindings phase. This diff adds it to `Context`, which is what controls
all the `Step` computations (including bindings).

Reviewed By: yangdanny97

Differential Revision: D74441068

fbshipit-source-id: 4f96f2c0e751ea8b849f425403a5503f9b10c3cb