Stress test CheckAndReturnAny mode
Summary:
There are a number of cases I didn't think about in advance of implementing
the mode to check function bodies but infer Any as the return type:
- Making sure yields and returns are all checked
- Making sure structural restrictions, like no return from async generator, are checked
- Making sure attribute types are still inferred

All of these work out-of-the-box, but we should have regression tests for them;
the test fixtures for attributes will also prove useful to reuse when
implementing the Skip mode because attribute handling is the trickiest aspect of
skip mode.

Note that in the process of this I noticed some badly worded error messages on
yield type errors, I left a TODO so I remember to come back to this eventually.

Reviewed By: yangdanny97

Differential Revision: D74441078

fbshipit-source-id: c698eeb003f7d0119ae7f3fa2ea367755b874c26